# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API backend para operacao de comandas.",
        "title": "VRComandaAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/atendentes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Atendente"
                ],
                "summary": "Listar atendentes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da loja",
                        "name": "id_loja",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Codigo",
                        "name": "codigo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nome",
                        "name": "nome",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Ativo",
                        "name": "ativo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/comandas": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comanda"
                ],
                "summary": "Listar comandas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da loja",
                        "name": "id_loja",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Numero da comanda",
                        "name": "comanda",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Numero de identificacao",
                        "name": "numero_identificacao",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Ativo",
                        "name": "ativo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/lancamentos": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lancamento"
                ],
                "summary": "Listar lancamentos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID comanda",
                        "name": "id_comanda",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID mesa",
                        "name": "id_mesa",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID atendente",
                        "name": "id_atendente",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "DataHora",
                        "name": "dataHora",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Finalizado",
                        "name": "finalizado",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lancamento"
                ],
                "summary": "Criar lancamento",
                "parameters": [
                    {
                        "description": "Lancamento",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lancamento.CreateLancamentoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/lancamentos/itens": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lancamento"
                ],
                "summary": "Listar itens por comanda",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID comanda",
                        "name": "id_comanda",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lancamento"
                ],
                "summary": "Criar itens de lancamento em lote",
                "parameters": [
                    {
                        "description": "Itens",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lancamento.CreateItemsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/lancamentos/itens/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lancamento"
                ],
                "summary": "Editar item de lancamento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lancamento.UpdateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/lancamentos/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lancamento"
                ],
                "summary": "Editar lancamento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lancamento",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lancamento.CreateLancamentoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mesas": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mesa"
                ],
                "summary": "Listar mesas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da loja",
                        "name": "id_loja",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Numero da mesa",
                        "name": "mesa",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Ativo",
                        "name": "ativo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "lancamento.CreateItemRequest": {
            "type": "object",
            "required": [
                "id_atendente",
                "id_lancamentocomanda",
                "id_produto",
                "quantidade",
                "sequencia"
            ],
            "properties": {
                "codigobarras": {
                    "type": "string"
                },
                "id_atendente": {
                    "type": "integer"
                },
                "id_lancamentocomanda": {
                    "type": "integer"
                },
                "id_produto": {
                    "type": "integer"
                },
                "id_situacao": {
                    "type": "integer"
                },
                "precovenda": {
                    "type": "number"
                },
                "quantidade": {
                    "type": "number"
                },
                "sequencia": {
                    "type": "integer"
                }
            }
        },
        "lancamento.CreateItemsRequest": {
            "type": "object",
            "required": [
                "itens"
            ],
            "properties": {
                "itens": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/lancamento.CreateItemRequest"
                    }
                }
            }
        },
        "lancamento.CreateLancamentoRequest": {
            "type": "object",
            "properties": {
                "dataHora": {
                    "type": "string"
                },
                "finalizado": {
                    "type": "boolean"
                },
                "id_atendente": {
                    "type": "integer"
                },
                "id_comanda": {
                    "type": "integer"
                },
                "id_loja": {
                    "type": "integer"
                },
                "id_mesa": {
                    "type": "integer"
                },
                "observacao": {
                    "type": "string"
                }
            }
        },
        "lancamento.UpdateItemRequest": {
            "type": "object",
            "properties": {
                "cancelado": {
                    "type": "boolean"
                },
                "quantidade": {
                    "type": "number"
                }
            }
        }
    }
}